liked cinema context, refactoring of showtimes component
